Update to bin script and moved read atf.
